backend/requirements.txt
# Healio Backend — Production Requirements
# Generated for Cloud Run deployment (python:3.11-slim)

# fastapi>=0.100.0
uvicorn>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.1.0
google-cloud-aiplatform>=1.50.0
google-api-core>=2.10.0
google-cloud-speech>=2.20.0
firebase-admin>=6.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
Pillow>=10.0.0
websockets>=11.0.0
python-socketio>=5.9.0
python-dotenv>=1.0.0
